add func save server to database; reformat/clean code

+++ iosApp/iosApp/Views/ServerDetailsView.swift
@@ -3,7 +3,7 @@ import shared
 
 struct ServerDetailsView: View {
     
-    @State private var serverId: Int = 0
+    @State private var serverId: Int64 = 0
     
     var server: Server
